+++ release-notes.md
@@ -4,6 +4,7 @@ Jupiter release notes
 ### 2017-05-06:
 
 - 存在addAndGet场景时, 不使用jupiter自带的Atomic*FieldUpdater, 使用jdk默认的实现, 1.8后续版本jdk使用了intrinsics后更快
+- 移除org.jupiter.rpc.model.metadata.ServiceWrapper#connCount, connCount修正为JServer粒度的参数
 
 ### 2017-04-18: version 1.2.4